Update README.md
The text Diccionario grande de la lengua de Michoacán (author unknown), which is the main referent for the vocabulary used in this project, was transcribed and edited by Dr. J. Benedict Warren, for its publication from Fimax Publicistas, Morelia, Michoacán, Mexico, in 1991, in two volumes. Vol. I, Español-Tarasco, and Vol. 2, Tarasco-Español, with an introduction by the same editor. Although there is no notice concerning the number of voices included in volume II, it is estimated that it includes approximately 25,000 entries in p’urhépecha (Later in the project, Vol. 1, which includes approximately the same number of entries, will serve as a counterpart for the translations). As this number of entries was considered enough for the first stage of the project, it was selected as the basis for the elaboration of the first version of the translator.

The Diccionario grande de la lengua de Michoacán was one of the reference works for the “Proyecto Gilberti” (Gilberti Project), which was part of the Ph. D. program in Ethnic Studies at the Center for the Study of Traditions (Centro de Estudios de las Tradiciones) of El Colegio de Michoacán. With the help of this and other available dictionaries, the transcription, translation and analysis of XVII-XIX c. texts, was done by native p’urhépecha-speakers for their Ph. D. theses. This means that once the project of the automatic translator is sufficiently advanced, it will be possible to incorporate the vocabulary of all transcribed texts, and that the translations made for the Ph. D. theses will be useful in the corroboration of the automatic translator’s results.

The electronic version of the Diccionario grande de la lengua de Michoacán was done by Dr. J. Benedict Warren (+) as the initial transcription before its publication.
...